requirements.txt
py3-validate-email==1.0.4